Refactoring files for better organization
I have refactored the componenets to hold a Home.js file that will consist of all the components required to be displayed on that component. Earlier this was done in the respective routes folder and I was caling individual component which eventually became messy as the size and the requirement of the page increased.